+++ CHANGELOG.rst
@@ -17,6 +17,8 @@ Version 0.13.0
   cannot have a storage section ``[storage foo]`` and a pair ``[pair foo]`` in
   your config, they have to have different names. This is done such that
   console output is always unambigous. See :gh:`459`.
+- Custom commands can now be used for conflict resolution during sync. See
+  :gh:`127`.
 
 Version 0.12.1
 ==============